# DomainIntelWeb/openapi.json
{
  "components": {
    "schemas": {
      "ArchiveState": {
        "properties": {
          "archived_to": {
            "title": "Archived To",
            "type": "string"
          }
        },
        "required": [
          "archived_to"
        ],
        "title": "ArchiveState",
        "type": "object"
      },
...
      "AutomationState": {
        "properties": {
          "email_delivery": {
            "default": false,
            "title": "Email Delivery",
            "type": "boolean"
          },
          "schedules": {
            "items": {
              "$ref": "#/components/schemas/ScheduleState"
            },
            "title": "Schedules",
            "type": "array"
          }
        },
        "required": [
          "schedules"
        ],
        "title": "AutomationState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IndustryCreate": {
        "properties": {
          "folder": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "default": "",
            "maxLength": 120,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder"
        ],
        "title": "IndustryCreate",
        "type": "object"
      },
      "IndustryMutationState": {
        "properties": {
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder",
          "name"
        ],
        "title": "IndustryMutationState",
        "type": "object"
      },
      "IndustryRename": {
        "properties": {
          "folder": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "default": "",
            "maxLength": 120,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "folder"
        ],
        "title": "IndustryRename",
        "type": "object"
      },
      "IndustryState": {
        "additionalProperties": true,
        "properties": {
          "folder": {
            "title": "Folder",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "periodic_enabled": {
            "default": false,
            "title": "Periodic Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "folder",
          "name"
        ],
        "title": "IndustryState",
        "type": "object"
      },
      "JobAccepted": {
        "properties": {
          "action": {
            "default": "",
            "title": "Action",
            "type": "string"
          },
          "email_delivery": {
            "default": false,
            "title": "Email Delivery",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
...
      "ScheduleState": {
        "properties": {
          "action": {
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "quarterly"
            ],
            "title": "Action",
            "type": "string"
          },
          "attempted_period_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempted Period Key"
          },
          "catch_up": {
            "title": "Catch Up",
            "type": "boolean"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "last_artifact_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Artifact Path"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_job_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Job Run Id"
          },
          "last_period_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Period Key"
          },
          "last_success_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "local_time": {
            "title": "Local Time",
            "type": "string"
          },
          "monthday": {
            "title": "Monthday",
            "type": "integer"
          },
          "next_run_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run At"
          },
          "pipeline_mode": {
            "default": "generate",
            "enum": [
              "aggregate",
              "generate"
            ],
            "title": "Pipeline Mode",
            "type": "string"
          },
          "provider": {
            "default": "codex",
            "title": "Provider",
            "type": "string"
          },
          "retry_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After"
          },
          "retry_count": {
            "default": 0,
            "title": "Retry Count",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "weekday": {
            "title": "Weekday",
            "type": "integer"
          }
        },
        "required": [
          "action",
          "enabled",
          "local_time",
          "weekday",
          "monthday",
          "timezone",
          "catch_up"
        ],
        "title": "ScheduleState",
        "type": "object"
      },
      "ScheduleUpdate": {
        "properties": {
          "catch_up": {
            "default": true,
            "title": "Catch Up",
            "type": "boolean"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "local_time": {
            "default": "08:00",
            "pattern": "^\\d{2}:\\d{2}$",
            "title": "Local Time",
            "type": "string"
          },
          "monthday": {
            "default": 1,
            "maximum": 28.0,
            "minimum": 1.0,
            "title": "Monthday",
            "type": "integer"
          },
          "pipeline_mode": {
            "default": "generate",
            "enum": [
              "aggregate",
              "generate"
            ],
            "title": "Pipeline Mode",
            "type": "string"
          },
          "provider": {
            "default": "codex",
            "enum": [
              "codex",
              "openai"
            ],
            "title": "Provider",
            "type": "string"
          },
          "timezone": {
            "default": "Asia/Shanghai",
            "maxLength": 100,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          },
          "weekday": {
            "default": 0,
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Weekday",
            "type": "integer"
          }
        },
        "title": "ScheduleUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "IntDog Local API",
    "version": "4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/artifact": {
      "get": {
        "operationId": "artifact_api_artifact_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 4000,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifact",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/industries": {
      "get": {
        "operationId": "industries_api_industries_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IndustryState"
                  },
                  "title": "Response Industries Api Industries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Industries",
        "tags": [
          "industries"
        ]
      },
      "post": {
        "operationId": "create_industry_api_industries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndustryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryMutationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Industry",
        "tags": [
          "industries"
        ]
      }
    },
    "/api/industries/{folder}": {
      "delete": {
        "operationId": "archive_industry_api_industries__folder__delete",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Industry",
        "tags": [
          "industries"
        ]
      },
      "patch": {
        "operationId": "rename_industry_api_industries__folder__patch",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndustryRename"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryMutationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Industry",
        "tags": [
          "industries"
        ]
      }
    },
    "/api/industries/{folder}/automation": {
      "get": {
        "operationId": "schedules_api_industries__folder__automation_get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedules",
        "tags": [
          "automation"
        ]
      }
    },
    "/api/industries/{folder}/automation/{action}": {
      "put": {
        "operationId": "configure_api_industries__folder__automation__action__put",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "enum": [
                "daily",
                "weekly",
                "monthly",
                "quarterly"
              ],
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Configure",
        "tags": [
          "automation"
        ]
      }
    },
    "/api/industries/{folder}/automation/{action}/run": {
      "post": {
        "operationId": "run_now_api_industries__folder__automation__action__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "enum": [
                "daily",
                "weekly",
                "monthly",
                "quarterly"
              ],
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Now",
        "tags": [
          "automation"
        ]
      }
    },
...
  }
}